divide to UserModel and ReviewModel
