+++ app/src/main/java/com/example/finalproject/model/FbReviewModel.java
@@ -18,7 +18,7 @@ public class FbReviewModel {
         db.setFirestoreSettings(settings);
     }
 
-    public void getAllReviews(Model.GetAllReviewsListener callback) {
+    public void getAllReviews(ReviewModel.GetAllReviewsListener callback) {
         db.collection("reviews").get().addOnCompleteListener((task) -> {
             ArrayList<Review> list = new ArrayList<>();
             if(task.isSuccessful()) {
@@ -33,7 +33,7 @@ public class FbReviewModel {
         });
     }
 
-    public void getUserReviewsSince(Long since, String user, Model.GetUserReviewsListener callback) {
+    public void getUserReviewsSince(Long since, String user, ReviewModel.GetUserReviewsListener callback) {
         db.collection("reviews").whereEqualTo("UserId", user).whereGreaterThanOrEqualTo("lastUpdated", new Timestamp(since,0)).get().addOnCompleteListener((task) -> {
             ArrayList<Review> list = new ArrayList<>();
             if(task.isSuccessful()) {
@@ -47,7 +47,7 @@ public class FbReviewModel {
         });
     }
 
-    public void addReview(Review review, Model.AddReviewListener listener) {
+    public void addReview(Review review, ReviewModel.AddReviewListener listener) {
         db.collection("reviews").document(review.getDocId()).set(review.toJson())
                 .addOnCompleteListener((task) -> {
 
@@ -55,7 +55,7 @@ public class FbReviewModel {
                 });
     }
 
-    public void updateReview(Review review, Model.UpdateReviewListener listener) {
+    public void updateReview(Review review, ReviewModel.UpdateReviewListener listener) {
         db.collection("reviews").document(review.getDocId()).update(review.toJson())
                 .addOnCompleteListener(task ->{
             listener.onComplete();
